all builds now with no warnings

+++ lib_mic_array/api/mic_array.h
@@ -2,7 +2,6 @@
 #define MIC_ARRAY_H_
 
 #include <stdint.h>
-
 #include "defines.h"
 #include "frame.h"
 
@@ -76,7 +75,7 @@ void hires_delay(
         streaming chanend c_4x_pdm_mic_1,
         streaming chanend c_sync,
         hires_delay_config * unsafe config,
-        unsigned long long * unsafe p_shared_memory_array);
+        int64_t * unsafe p_shared_memory_array);
 /*
  * Four Channel Deciamtor config structure.
  */
